requirements.txt
flask
twilio
google-genai
python-dotenv
pypdf
langchain-google-genai
